weather_data_pipeline/requirements.txt
# File: requirements.txt

apache-airflow==2.5.1
pandas==1.4.4
requests==2.28.1
pyyaml==6.0